Moved teaminfo and analysistypes definition files to lib folder, updated shruti's title based on request
